[Feature] login & logout initial index script (#7)
This pull request includes the addition of basic HTML structure for the
login and logout pages in the `public` directory. The changes primarily
involve setting up the HTML skeleton, linking to the CSS stylesheet, and
including a favicon.

HTML structure setup:

*
[`public/login/index.php`](diffhunk://#diff-6703a2b5d2e4fd6c09ba562f619c268e8bd7ca66565a848e6109d27aafa3b4daR1-R22):
Added basic HTML structure for the login page, including a link to the
stylesheet and favicon.
*
[`public/logout/index.php`](diffhunk://#diff-c53b482a65ac5860e242412c40b4610ea06f2c2bebf50c40eb2e165a7a195fc0R1-R22):
Added basic HTML structure for the logout page, including a link to the
stylesheet and favicon.