[docs] create contributing.md (#9)
This pull request consolidates the contribution guidelines into the
`CONTRIBUTING.md` file and updates references in the `README.md` to
point to this new location. The changes ensure that all contributors
follow a unified set of guidelines for documentation, coding
conventions, and commit messages.

Documentation updates:

*
[`CONTRIBUTING.md`](diffhunk://#diff-eca12c0a30e25b4b46522ebf89465a03ba72a03f540796c979137931d8f92055R1-R20):
Added comprehensive guidelines for documentation, coding conventions,
testing, commit messages, pull requests, and review and approval
process.

References update:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L18-R18):
Updated the link to the contribution guidelines to point to
`CONTRIBUTING.md` in the table of contents.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R129-L152):
Removed the detailed contribution guidelines and replaced them with a
link to `CONTRIBUTING.md`.

---------

Co-authored-by: UzziKyle <uzzielkyleynciong0824@gmail.com>
Co-authored-by: Uzziel Kyle <101367055+uzzielkyle@users.noreply.github.com>